Move atif converter utility from eval to core
